+++ backend/canisters/community/impl/src/lifecycle/post_upgrade.rs
@@ -1,6 +1,7 @@
+use crate::jobs::import_groups::finalize_group_import;
 use crate::lifecycle::{init_env, init_state, UPGRADE_BUFFER_SIZE};
 use crate::memory::get_upgrades_memory;
-use crate::Data;
+use crate::{read_state, Data};
 use canister_logger::LogEntry;
 use canister_tracing_macros::trace;
 use community_canister::post_upgrade::Args;
@@ -21,6 +22,12 @@ fn post_upgrade(args: Args) {
     canister_logger::init_with_logs(data.test_mode, logs, traces);
 
     init_state(env, data, args.wasm_version);
+
+    let completed_imports = read_state(|state| state.data.groups_being_imported.completed_imports());
+
+    for group_id in completed_imports {
+        finalize_group_import(group_id);
+    }
 
     info!(version = %args.wasm_version, "Post-upgrade complete");
 }